style: miscellaneous changes to site
