+++ data/timeline.json
@@ -7,7 +7,7 @@
             "date": "Jan - Mar",
             "topics" : [
                 {
-                    "topic" : "Pearl became the \"Agent App Store\"",
+                    "topic" : "Pearl became the \"AI Agent App Store\"",
                     "content" : 
                         "- Pearl now offers users a growing range of AI agents in one app. \n- These user-owned autonomous AI agents work for you and earn OLAS staking rewards, all in a user-friendly, open-source desktop app."
                     ,
@@ -34,7 +34,7 @@
                 },
                 {
                     "topic" : "Olas Accelerator Launched",
-                    "content" : "- Grant + mentorship program for teams building for Pearl's Agent App Store \n- 4 teams accepted: RateRover, [@Celo](https://x.com/Celo), [@BioProtocol](https://x.com/BioProtocol), [@TrendMoon](https://x.com/0xTrendMoon) \n- 40+ applicants, 6 spots remain open",
+                    "content" : "- Grant + mentorship program for teams building for Pearl's AI Agent App Store \n- 4 teams accepted: RateRover, [@Celo](https://x.com/Celo), [@BioProtocol](https://x.com/BioProtocol), [@TrendMoon](https://x.com/0xTrendMoon) \n- 40+ applicants, 6 spots remain open",
                     "category": ["Build", "Key milestones"]
                 },
                 {
