+++ pages/accelerator.jsx
@@ -6,7 +6,7 @@ const AcceleratorPage = () => (
   <PageWrapper>
     <Meta
       pageTitle="Olas Accelerator"
-      description="Build AI agents for Pearl, the Agent App Store, and earn rewards. Create AI agents and get paid in the growing AI and crypto space."
+      description="Build AI agents for Pearl, the AI Agent App Store, and earn rewards. Create AI agents and get paid in the growing AI and crypto space."
     />
     <Accelerator />
   </PageWrapper>
